Sponsor: add hardware check
